Show offline mirrors in the document browser
